+++ app/src/main/java/com/instacart/library/sample/SampleActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.instacart.library.truetime.time.TrueTime
+import com.instacart.library.truetime.legacy.TrueTime
 import kotlinx.android.synthetic.main.activity_sample.*
 import java.text.SimpleDateFormat
 import java.util.Date
