+++ app/src/main/java/com/instacart/library/sample/SampleRxActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.instacart.library.truetime.time.TrueTimeRx
+import com.instacart.library.truetime.legacy.TrueTimeRx
 import kotlinx.android.synthetic.main.activity_sample.*
 import java.text.SimpleDateFormat
 import java.util.Date
