+++ library/src/main/java/com/instacart/library/truetime/legacy/TrueLog.kt
@@ -1,4 +1,4 @@
-package com.instacart.library.truetime
+package com.instacart.library.truetime.legacy
 
 import android.util.Log
