+++ library/src/main/java/com/instacart/library/truetime/legacy/TrueTime.java
@@ -1,13 +1,9 @@
-package com.instacart.library.truetime.time;
-
+package com.instacart.library.truetime.legacy;
 
 import android.content.Context;
 import android.os.SystemClock;
 import com.instacart.library.truetime.cache.CacheInterface;
-import com.instacart.library.truetime.cache.DiskCacheClient;
 import com.instacart.library.truetime.cache.SharedPreferenceCacheImpl;
-import com.instacart.library.truetime.TrueLog;
-import com.instacart.library.truetime.sntp.SntpClient;
 import java.io.IOException;
 import java.util.Date;
 import java.util.Locale;
